Charles Palmer's review of PR #141 (comment rewrites, defines-to-top,
ae_monitor.py fix, light_sensor.md, and his review notes), plus the
development-reports thread: review responses, bench validation evidence with
serial logs, the reviewer worktree workflow and the dual-image build guide.
Build outputs restored to their canonical versions (#169).

+++ EPII_CM55M_APP_S/app/ww_projects/ww500_md/exif_builder.h
@@ -94,7 +94,7 @@ typedef struct {
     const char    *maker_note;      /* AE register CSV string, or NULL */
     const char    *camera_model;    /* EXIF Model, e.g. "WW500 RP3"; NULL -> "WW500" */
     const char    *software;        /* EXIF Software: firmware build string, or NULL to omit */
-    uint16_t       flash_fired;     /* EXIF Flash: 1 = flash/IR illumination used for this capture, 0 = not */
+    uint8_t       flash_fired;     /* EXIF Flash: 0 if flash is inactive. Otherwise 1 (visible) or 2 (IR)  */
 } ExifInput_t;
 
 /*************************************** Public Functions *******************************************/
